Change the Riccati solver to a single-precision floating-point number.

+++ testbench/riccati_solver/riccati_solver.h
@@ -1,6 +1,6 @@
 #ifndef RICCATI_SOLVER_H_
 #define RICCATI_SOLVER_H_
 
-void solveRiccatiIteration(double A[3][3], double B[3][2], double Q[3][3], double R[2][2]);
+void solveRiccatiIteration(float A[3][3], float B[3][2], float Q[3][3], float R[2][2]);
 
 #endif
